HTTPS

+++ NodeJS/conFusionServer/routes/dishRouter.js
@@ -21,6 +21,7 @@ dishRouter.route('/')
     .catch((err) => next(err));
 })
 .post(authenticate.verifyUser, (req,res,next) => {
+    authenticate.verifyAdmin(req, res, next);
     Dishes.create(req.body)
     .then((dish) => {
         console.log('Dish Created ', dish);
@@ -31,11 +32,13 @@ dishRouter.route('/')
     .catch((err) => next(err));
 })
 .put(authenticate.verifyUser, (req,res,next) => {
+    authenticate.verifyAdmin(req, res, next);
     res.statusCode = 403;
     res.end('PUT operation not supported on /dishes');
 })
 
 .delete(authenticate.verifyUser, (req,res,next) => {
+    authenticate.verifyAdmin(req, res, next);
     Dishes.remove({})
     .then((resp) => {
         res.statusCode = 200;
@@ -57,10 +60,12 @@ dishRouter.route('/:dishId')
     .catch((err) => next(err));
 })
 .post(authenticate.verifyUser, (req, res, next) => {
+    authenticate.verifyAdmin(req, res, next);
     res.statusCode = 403;
     res.end('POST operation not supported on /dishes/'+ req.params.dishId);
 })
 .put(authenticate.verifyUser, (req, res, next) => {
+    authenticate.verifyAdmin(req, res, next);
     Dishes.findByIdAndUpdate(req.params.dishId, {
         $set: req.body
     }, { new: true })
@@ -72,6 +77,7 @@ dishRouter.route('/:dishId')
     .catch((err) => next(err));
 })
 .delete(authenticate.verifyUser, (req, res, next) => {
+    authenticate.verifyAdmin(req, res, next);
     Dishes.findByIdAndRemove(req.params.dishId)
     .then((resp) => {
         res.statusCode = 200;
@@ -130,6 +136,7 @@ dishRouter.route('/:dishId/comments')
 })
 
 .delete(authenticate.verifyUser, (req,res,next) => {
+    authenticate.verifyAdmin(req, res, next);
     Dishes.findById(req.params.dishId)
     .then((dish) => {
         if(dish != null) {
@@ -181,6 +188,11 @@ dishRouter.route('/:dishId/comments/:commentId')
     '/comments/' + req.params.commentId);
 })
 .put(authenticate.verifyUser, (req, res, next) => {
+    if(!req.user._id.equals(req.params.comment.author)){
+        err = new Error("You cannot update someone else's comment!");
+        err.status = 403;
+        return next(err);
+    }
     Dishes.findById(req.params.dishId)
     .then((dish) => {
         if(dish != null && dish.comments.id(req.params.commentId)  != null) {
@@ -215,6 +227,11 @@ dishRouter.route('/:dishId/comments/:commentId')
     .catch((err) => next(err));
 })
 .delete(authenticate.verifyUser, (req, res, next) => {
+    if(!req.user._id.equals(req.params.comment.author)){
+        err = new Error("You cannot delete someone else's comment!");
+        err.status = 403;
+        return next(err);
+    }
     Dishes.findById(req.params.dishId)
     .then((dish) => {
         if(dish != null && dish.comments.id(req.params.commentId)  != null) {
