+++ NodeJS/conFusionServer/routes/leaderRouter.js
@@ -20,6 +20,7 @@ leaderRouter.route('/')
     .catch((err) => next(err));
 })
 .post(authenticate.verifyUser, (req,res,next) => {
+    authenticate.verifyAdmin(req, res, next);
     Leaders.create(req.body)
     .then((leader) => {
         console.log('Leader Created ', leader);
@@ -30,11 +31,13 @@ leaderRouter.route('/')
     .catch((err) => next(err));
 })
 .put(authenticate.verifyUser, (req,res,next) => {
+    authenticate.verifyAdmin(req, res, next);
     res.statusCode = 403;
     res.end('PUT operation not supported on /leaders');
 })
 
 .delete(authenticate.verifyUser, (req,res,next) => {
+    authenticate.verifyAdmin(req, res, next);
     Leaders.remove({})
     .then((resp) => {
         res.statusCode = 200;
@@ -55,10 +58,12 @@ leaderRouter.route('/:leaderId')
     .catch((err) => next(err));
 })
 .post(authenticate.verifyUser, (req, res, next) => {
+    authenticate.verifyAdmin(req, res, next);
     res.statusCode = 403;
     res.end('POST operation not supported on /leaders/'+ req.params.leaderId);
 })
 .put(authenticate.verifyUser, (req, res, next) => {
+    authenticate.verifyAdmin(req, res, next);
     Leaders.findByIdAndUpdate(req.params.leaderId, {
         $set: req.body
     }, { new: true })
@@ -70,6 +75,7 @@ leaderRouter.route('/:leaderId')
     .catch((err) => next(err));
 })
 .delete(authenticate.verifyUser, (req, res, next) => {
+    authenticate.verifyAdmin(req, res, next);
     Leaders.findByIdAndRemove(req.params.leaderId)
     .then((resp) => {
         res.statusCode = 200;
