+++ NodeJS/conFusionServer/routes/promoRouter.js
@@ -20,6 +20,7 @@ promoRouter.route('/')
     .catch((err) => next(err));
 })
 .post(authenticate.verifyUser, (req,res,next) => {
+    authenticate.verifyAdmin(req, res, next);
     Promotions.create(req.body)
     .then((promo) => {
         console.log('Promotion Created ', promo);
@@ -30,11 +31,13 @@ promoRouter.route('/')
     .catch((err) => next(err));
 })
 .put(authenticate.verifyUser, (req,res,next) => {
+    authenticate.verifyAdmin(req, res, next);
     res.statusCode = 403;
     res.end('PUT operation not supported on /promotions');
 })
 
 .delete(authenticate.verifyUser, (req,res,next) => {
+    authenticate.verifyAdmin(req, res, next);
     Promotions.remove({})
     .then((resp) => {
         res.statusCode = 200;
@@ -55,10 +58,12 @@ promoRouter.route('/:promoId')
     .catch((err) => next(err));
 })
 .post(authenticate.verifyUser, (req, res, next) => {
+    authenticate.verifyAdmin(req, res, next);
     res.statusCode = 403;
     res.end('POST operation not supported on /promotions/'+ req.params.promoId);
 })
 .put(authenticate.verifyUser, (req, res, next) => {
+    authenticate.verifyAdmin(req, res, next);
     Promotions.findByIdAndUpdate(req.params.promoId, {
         $set: req.body
     }, { new: true })
@@ -70,6 +75,7 @@ promoRouter.route('/:promoId')
     .catch((err) => next(err));
 })
 .delete(authenticate.verifyUser, (req, res, next) => {
+    authenticate.verifyAdmin(req, res, next);
     Promotions.findByIdAndRemove(req.params.promoId)
     .then((resp) => {
         res.statusCode = 200;
